Moved mockers clsses out of AbstractIntegrationTest class
